+++ R/global_wlc.R
@@ -2,11 +2,13 @@
 #'
 #' @param cost_surface cost surface
 #'
-#' @param p risk
-#'
 #' @param constrains Boolean vector of adjacent cells not traversable
 #'
+#' @param fun An R function to apply to the standardised transition values
+#'
 #' @param global_weight Importance of factor
+#'
+#' @param neighbours xxx
 #'
 #' @return Global Weighted Linear Combination
 #'
@@ -14,19 +16,14 @@
 #'
 #' @export
 
-global_wlc <- function(cost_surface, p = 1, constrains = NULL, global_weight) {
-
-  if (p <= 1) {stop("p must be equal or greater than 1")}
-
-  # if(!inherits(constrains, "logical")) {stop("constraints must be a logical vector")}
-
-  if(global_weight > 1 | global_weight <= 0) {stop("global weight must be greater than 0 and less than or equal to 1")}
+global_wlc <- function(cost_surface, constrains = NULL, global_weight, neighbours, fun = NULL) {
 
   # remove explicit zeroes from the sparse matrix. Ensures scaling is not impacted by zero values
   cost_surface@transitionMatrix <- Matrix::drop0(cost_surface@transitionMatrix)
 
   # rescale to between 0 and 1
-  cost_surface <- rescale_cs_global(cost_surface = cost_surface, p = p, constrains = constrains)
+  cost_surface <- rescale_cs_global(cost_surface = cost_surface, constrains = constrains, neighbours = neighbours, fun = fun)
+
   # multiply by user-supplied criterion global weight
   cost_surface <- cost_surface * global_weight