+++ R/rescale_cs_global.R
@@ -2,9 +2,11 @@
 #'
 #' @param cost_surface cost surface
 #'
-#' @param p risk
+#' @param constrains Logical vector of adjacent cells not traversable
 #'
-#' @param constrains Boolean vector of adjacent cells not traversable
+#' @param neighbours xxx
+#'
+#' @param fun An R function to apply to the standardised transition values
 #'
 #' @return cost suface standardised to a maximum value of 1 and minimum value relative to 0
 #'
@@ -12,15 +14,16 @@
 #'
 #' @export
 
-rescale_cs_global <- function(cost_surface, p = 1, constrains = NULL) {
-
-  if (p <= 1) {stop("p must be equal or greater than 1")}
+rescale_cs_global <- function(cost_surface, constrains = NULL, neighbours, fun = NULL) {
 
   cost_surface@transitionMatrix <- Matrix::drop0(cost_surface@transitionMatrix)
 
-  cs_adj <- gdistance::adjacencyFromTransition(cost_surface)
+  rast_cs <- suppressWarnings(raster::raster(cost_surface))
 
-  if (!is.null(constrains)) {
+  cs_adj <- raster::adjacent(rast_cs, cells = 1:raster::ncell(rast_cs),
+                             pairs = TRUE, directions = neighbours)
+
+  if(inherits(constrains, "logical"))  {
     cost_surface[cs_adj][constrains] <- 0
   }
 
@@ -28,7 +31,11 @@ rescale_cs_global <- function(cost_surface, p = 1, constrains = NULL) {
 
   max_val <- base::max(cost_surface@transitionMatrix@x)
 
-  cost_surface@transitionMatrix@x <- (((c(0, cost_surface@transitionMatrix@x) - 0) / (max_val - 0)) ^ p)[-1]
+  cost_surface@transitionMatrix@x <- (((c(0, cost_surface@transitionMatrix@x) - 0) / (max_val - 0)))[-1]
+
+  if(inherits(fun, "function"))  {
+    cost_surface@transitionMatrix@x <- fun(cost_surface@transitionMatrix@x)
+  }
 
   return(cost_surface)
 }