+++ NEWS.md
@@ -1,3 +1,8 @@
+# 0.1.4 (17th June 2022)
+* Modified rescale_cs_global to allow for user-supplied transformations of transition values via the fun argument
+* Added fun argument to  global_wlc
+* Added change_tolerance function for ease-of-use when modifying/testing tolerance threshold values
+
 # 0.1.3 (11th January 2022)
 * Implemented Local and Global Weighted Linear Combination
   - rescale_cs_global